refactor(gateway): DaemonSets on a labelled node, and uninstall the old daemons
Three changes, all about the transports actually owning the box rather than
appearing to.

The k8s modules take the plain names — xray.ts, hysteria.ts, tailscale.ts —
and the SSH ones become *-systemd.ts. Nothing else here is named after the
Kubernetes primitive it happens to use, and these are how the transports are
deployed now; the systemd path is the one that needs qualifying, since it is
what an edge has and what a gateway without k3s falls back to.

DaemonSets rather than Deployments with replicas: 1. With hostNetwork two
replicas can never coexist on a node, so "exactly one per matching node" is
what these are, and `strategy: Recreate` was working around an update that
cannot succeed by construction. maxUnavailable 1 / maxSurge 0 is spelled out
because the ordering is load-bearing on an exclusive port. A nodeSelector on
jaritanet.dev/vpn-entry keeps them off `lady` when it joins, and makes moving
an entry a label rather than a module.

The teardown now uninstalls. Masking leaves the binary and unit sitting there
for a package upgrade to resurrect onto a port a pod is holding, so xray and
hysteria go through their vendor uninstalls (covering the templated
hysteria-server@ instances the alt ports run as) and unbound is purged.
tailscale is removed, never purged: the pod inherits the node's tailnet
identity from /var/lib/tailscale and purge deletes it, which would put the
cluster up and out of reach. The command ends by asserting the end state —
no such binary present, nothing outside kubepods on tcp/443, udp/443,
udp/3478, udp/4500 or 127.0.0.1:53 — because a pod that loses a port race
still reports Running.

Also: node sysctls for the UDP buffers quic-go asks for (Ubuntu caps
rmem_max at 208KB, hysteria documents 16MB) beside the existing BBR tuning,
since hostNetwork means the node's values are the ones in force; CPU
requests with no CPU limits, because CFS throttling on a tunnel is latency
spikes rather than a graceful slowdown; and unbound's memory limit raised off
192Mi, which was its declared cache size exactly, with nothing left for the
process holding it.

Checked rather than assumed: xray 26.3.27 is a plain go1.26.1 linux/amd64
build, so AES-NI and AVX2 come from runtime CPU detection and seccomp, which
filters syscalls, cannot touch them; hysteria v2.10.0 serves clean under the
default seccomp profile with host networking.
