+++ docs/architecture.md
@@ -155,15 +155,36 @@ tailscaled and unbound are Deployments in it rather than systemd units installed
 over SSH. The daemons and their configuration are unchanged; what changes is who
 starts them and where their secrets come from.
 
-All four pods are `hostNetwork`, and that is a requirement rather than a
-convenience. Xray must own the host's real `:443` and see real client source
-addresses, and its `dest` is `127.0.0.1:8443` — Traefik's `hostPort`, which only
-means Traefik from inside the host's namespace. unbound answers `127.0.0.1:53`
-for clients dialling that loopback *through* the tunnel. tailscaled has to put
-`tailscale0` where the other two resolve their routes. All four use
-`strategy: Recreate`: a host port is exclusive, so a rolling update leaves the
-new pod Pending on a bind the old pod will not release until it is allowed to
-leave — the deadlock Traefik's `hostPort` already demonstrated.
+All four are `hostNetwork`, and that is a requirement rather than a convenience.
+Xray must own the host's real `:443` and see real client source addresses, and
+its `dest` is `127.0.0.1:8443` — Traefik's `hostPort`, which only means Traefik
+from inside the host's namespace. unbound answers `127.0.0.1:53` for clients
+dialling that loopback *through* the tunnel. tailscaled has to put `tailscale0`
+where the other two resolve their routes.
+
+That is also why they are **DaemonSets**. With `hostNetwork` two replicas can
+never coexist on a node — they would collide on the same ports — so "exactly one
+per matching node" is the shape, and the primitive enforces it instead of a
+Deployment's update strategy working around it. The update strategy still
+matters for the same reason: `maxUnavailable: 1` with `maxSurge: 0` (the
+DaemonSet default, written out because it is load-bearing) deletes the old pod
+before creating its replacement, which is the only order that can succeed on an
+exclusive port. Traefik sat 20 days on a stale pod for getting this wrong.
+
+They select on the `jaritanet.dev/vpn-entry` node label rather than a hostname,
+so serving an entry is a property of a node. `lady` joining the cluster does not
+make it a VPN entry, and making an edge one later is a label rather than another
+module. The gateway labels itself after k3s comes up; a missing label would be
+silent — the DaemonSet would just schedule nothing.
+
+Two things they need from the node rather than from their own spec, both because
+`hostNetwork` means the node's networking is theirs: BBR and `fq`, and
+`net.core.rmem_max`/`wmem_max` at 16MB. quic-go asks Hysteria2's UDP socket for
+a multi-megabyte receive buffer and the kernel silently caps it at `rmem_max`,
+which Ubuntu ships at 208KB — the symptom is dropped datagrams under load and a
+warning on the connection path, not at start-up. Both live in
+`createNetworkTuning`, which asserts the values took rather than only writing
+the file.
 
 The REALITY keypair, its shortId, every client UUID, the hy2 obfs and per-admin
 passwords and the hy2 self-signed cert are all generated by Pulumi and mounted
@@ -174,12 +195,23 @@ or rebuilt. The REALITY key is *derived* from a stored 32-byte seed rather than
 generated per run (`realityKeypair`), or every deploy would replace the inbound
 and invalidate every client profile.
 
-Migrating a gateway that was provisioned the old way needs its units stopped
-first: Pulumi dropping a remote command from the program runs nothing on the
-machine, so xray would still hold `:443` when the pod tries to bind it.
-`gateway-legacy-units` disables them, and every pod is ordered after it. It
-deliberately does not `tailscale logout` — the pod inherits `/var/lib/tailscale`
-and with it the node identity, so a logout would register a second machine.
+Migrating a gateway that was provisioned the old way needs its daemons
+**uninstalled**, not stopped: Pulumi dropping a remote command from the program
+runs nothing on the machine, and a merely stopped unit is one package upgrade or
+postinst away from taking a port back from the pod holding it.
+`gateway-legacy-units` runs the vendor uninstalls for xray and Hysteria2 (which
+covers the templated `hysteria-server@` instances the alt ports run as), purges
+unbound, and every pod is ordered after it. It finishes by asserting the end
+state — no such binary left, and no non-container process on tcp/443, udp/443,
+udp/3478, udp/4500 or `127.0.0.1:53` — because the failure it is preventing is
+invisible from `kubectl`: the pod that loses a port race still reports Running.
+
+tailscale is the exception: `apt-get remove`, never `purge`. The pod inherits
+this node's tailnet identity from `/var/lib/tailscale`, and purge deletes that
+directory, so the relay would rejoin as a new machine needing a fresh authkey —
+the cluster up but unreachable. The teardown copies the state aside, removes the
+package, restores it if the packaging took it anyway, and fails loudly if the
+node key did not survive.
 
 Edges keep the SSH path: they have no cluster, and one box per location running
 two daemons is not worth a control plane.
